+++ elftools/elf/structs.py
@@ -11,10 +11,9 @@ from ..construct import (
     UBInt8, UBInt16, UBInt32, UBInt64,
     ULInt8, ULInt16, ULInt32, ULInt64,
     SBInt32, SLInt32, SBInt64, SLInt64,
-    Struct, Array, Enum, Padding, BitStruct, BitField, Value, String, CString,
-    Rename, RepeatUntil, Adapter, Field
+    Struct, Array, Enum, Padding, BitStruct, BitField, Value, String, CString
     )
-
+from ..common.construct_utils import ULEB128
 from .enums import *
 
 
@@ -121,8 +120,7 @@ class ELFStructs(object):
         )
 
     def _create_leb128(self):
-        self.Elf_uleb128 = _ULEB128
-        self.Elf_sleb128 = _SLEB128
+        self.Elf_uleb128 = ULEB128
 
     def _create_ntbs(self):
         self.Elf_ntbs = CString
@@ -374,47 +372,3 @@ class ELFStructs(object):
                                         Enum(self.Elf_uleb128('tag'),
                                              **ENUM_ATTR_TAG_ARM)
         )
-
-
-def _LEB128_reader():
-    """ Read LEB128 variable-length data from the stream. The data is terminated
-        by a byte with 0 in its highest bit.
-    """
-    return RepeatUntil(
-                lambda obj, ctx: ord(obj) < 0x80,
-                Field(None, 1))
-
-
-class _ULEB128Adapter(Adapter):
-    """ An adapter for ULEB128, given a sequence of bytes in a sub-construct.
-    """
-    def _decode(self, obj, context):
-        value = 0
-        for b in reversed(obj):
-            value = (value << 7) + (ord(b) & 0x7F)
-        return value
-
-
-class _SLEB128Adapter(Adapter):
-    """ An adapter for SLEB128, given a sequence of bytes in a sub-construct.
-    """
-    def _decode(self, obj, context):
-        value = 0
-        for b in reversed(obj):
-            value = (value << 7) + (ord(b) & 0x7F)
-        if ord(obj[-1]) & 0x40:
-            # negative -> sign extend
-            value |= - (1 << (7 * len(obj)))
-        return value
-
-
-def _ULEB128(name):
-    """ A construct creator for ULEB128 encoding.
-    """
-    return Rename(name, _ULEB128Adapter(_LEB128_reader()))
-
-
-def _SLEB128(name):
-    """ A construct creator for SLEB128 encoding.
-    """
-    return Rename(name, _SLEB128Adapter(_LEB128_reader()))
